+++ NEWS.md
@@ -1,3 +1,5 @@
+# connectapi (development version)
+
 # connectapi development
 
 BREAKING: