+++ NEWS.md
@@ -1,3 +1,5 @@
+# connectapi (development version)
+
 # connectapi 0.1.0.9018
 
 - Add a `client$PATCH` verb